+++ CHANGELOG.md
@@ -11,7 +11,13 @@ All Notable changes to `digipolisgent/toerismevlaanderen-lodging` package.
 ### Changed
 
 * VG-1660: Replaced the single phone number, email address, website address by
-           collections.
+  collections.
+* VG-1660: Replaced star ratings by rating interface: the returned rating value
+  can also be a category rating or can be empty.
+
+### Fixed
+
+* VG-1660: Fixed the lodging details query.
 
 ## [0.1.0]